requirements.txt
Django==5.1.2
djangorestframework==3.15.2
drf-yasg==1.21.8
isort==5.13.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
